// src/sdk/entities/project/equipment/LightingFixture.ts
import { Moment } from "moment";
import { Entity } from "../../../contracts/Entity";
import { EntityType } from "../../../sdkconfig/EntityTypeId";
import { Cost } from "./Cost";
import { DatasheetEntry } from "./DatasheetEntry";
import { LampConfiguration } from "./LampConfiguration";
import { Measurement } from "./Measurement";

@EntityType("1c104b38-c9e2-4ca3-bde9-189ea6c20590")
export class LightingFixture extends Entity {
  Model: string = "";
  Description: string = "";
  Watts: number;
  Voltage: number;
  VoltageThru: number;
  Finish: string = "";
  Notes: string = "";
  InternalNotes: string = "";
  IsCustom: boolean;
  IsModified: boolean;
  Nickname: string = "";
  IsDimmable: boolean;
  Glass: string = "";
  IsAda: boolean;
  MaxDmxChannels: number;
  HasData: boolean;
  HasPower: boolean;
  CanUseData: boolean;
  CanUsePower: boolean;
  IsAC: boolean;
  CurrentCost: Cost = new Cost();
  Measurement: Measurement = new Measurement();
  LampConfiguration: LampConfiguration = new LampConfiguration();
  Rating: string = "";
  Connector: string = "";
  Class: string = "";
  Title: string = "";
  DatasheetAsOf: Moment;
  CanTakeGobo: boolean;
  CanTakeFilter: boolean;
  AccessDbid: string = "";
  ManufacturerId: string = "";
  CurrentListings: string[] = [];
  Datasheets: DatasheetEntry[] = [];
  Shade: string;
  ShadeMaterial: string;
  CreationDate: Moment;
  LightSource: string;
  IsInactive: boolean;
  ColorTempStart: Number;
  ColorTempEnd: Number;
}